requirements.txt
Pinata-python
web3
python-dotenv
wave
openai
requests
pydub
sounddevice
numpy
nltk
speechrecognition
#scipy
pyserial
#mutagen
#regex_spm
toml
pygame
keyboard
#eth-tester
#py-evm
oz-defender
torch
openai-whisper
whisper-mic
langchain
langchain-community
huggingface_hub
elevenlabs
simpleaichat
text-generation
transformers
google-search-results
numexpr
langchainhub
sentencepiece
jinja2